refactor: set user module to users infra
