--- src/modules/customers/infra/http/user.module.ts
@@ -1,24 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { RepositoryEnum } from '../../../../shared/enums/repositories';
-import { ServiceEnum } from 'src/shared/enums/services';
-import { TypeormUserRepository } from 'src/modules/users/infra/database/typeorm/repositories/typeorm-user.repository';
-import { TypeormUserModel } from 'src/modules/users/infra/database/typeorm/models/typeorm-user.model';
-import { UserService } from 'src/modules/users/application/services/user.service';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([TypeormUserModel])],
-  controllers: [],
-  providers: [
-    {
-      provide: RepositoryEnum.USER,
-      useClass: TypeormUserRepository,
-    },
-    {
-      provide: ServiceEnum.USER,
-      useClass: UserService,
-    },
-  ],
-  exports: [ServiceEnum.USER],
-})
-export class UserModule {}
