+++ src/modules/groups/infra/http/group.module.ts
@@ -6,8 +6,9 @@ import { TypeormGroupModel } from '../../infra/database/typeorm/models/typeorm-g
 import { TypeormGroupMemberModel } from '../../infra/database/typeorm/models/typeorm-group-member.model';
 import { TypeormGroupRepository } from '../../infra/database/typeorm/repositories/typeorm-group.repository';
 import { TypeormGroupMemberRepository } from '../../infra/database/typeorm/repositories/typeorm-group-member.repository';
-import { UserModule } from 'src/modules/customers/infra/http/user.module';
+
 import { CreateGroupUseCase } from '../../application/use-cases/create-group/create-group.use-case';
+import { UserModule } from 'src/modules/users/infra/http/user.module';
 
 @Module({
   imports: [
